Fit spOccupancy model

Fit spOccupancy model

Usage

fitSpOccu(
  site_data_year,
  visit_data_year,
  config,
  sp_code,
  spatial = FALSE,
  sp_sites,
  ...
)

Arguments

site_data_year

Occupancy site data for a single year and species (see ppl_create_site_visit)

visit_data_year

Occupancy visit data for a single year and species (see ppl_create_site_visit)

config

A list with pipeline configuration parameters (see configPipeline).

sp_code

SAFRING code of the species the pipeline is running for

spatial

Logical, indicating whether spatial random effects should be included in the model (TRUE) or not (FALSE, default)

sp_sites

Spatial object containing the pentads in site_data_year.

...

Other arguments that might be needed (e.g. for messages)

Value

Either a spOccupancy model fit or the integer 3, indicating that model fit failed.