Control of BIRDiE Model Fitting

Constructs control parameters for BIRDiE model fitting. All arguments have defaults.

Usage

birdie.ctrl(
  max_iter = 1000,
  accel = c("squarem", "anderson", "daarem", "none"),
  order = switch(match.arg(accel), none = 0L, anderson = -1L, daarem = -1L, squarem = 1L),
  anderson_restart = TRUE,
  abstol = 1e-06,
  reltol = 1e-06
)

Arguments

max_iter

The maximum number of EM iterations.

accel

The acceleration algorithm to use in doing EM. The default "squarem" is good for most purposes, though "anderson" may be faster when there are few parameters or very tight tolerances. "daarem" is an excellent choice as well that works across a range of problems, though it requires installing the small daarem package. "none" is not recommended unless other algorithms are running into numerical issues. See the references below for details on these schemes.

order

The order to use in the acceleration algorithm. Interpretation varies by algorithm. Can range from 1 (default) to 3 for SQUAREM and from 1 to the number of parameters for Anderson and DAAREM (default -1 allows the order to be determined by problem size).

anderson_restart

Whether to use restarts in Anderson acceleration.

abstol

The absolute tolerance used in checking convergence.

reltol

The relative tolerance used in checking convergence. Ignored if accel = "squarem" or "daarem".

Value

A list containing the control parameters.

References

Varadhan, R., & Roland, C. (2004). Squared extrapolation methods (SQUAREM): A new class of simple and efficient numerical schemes for accelerating the convergence of the EM algorithm.

Walker, H. F., & Ni, P. (2011). Anderson acceleration for fixed-point iterations. SIAM Journal on Numerical Analysis, 49(4), 1715-1735.

Henderson, N. C., & Varadhan, R. (2019). Damped Anderson acceleration with restarts and monotonicity control for accelerating EM and EM-like algorithms. Journal of Computational and Graphical Statistics, 28(4), 834-846.

Examples

birdie.ctrl(max_iter=1000)
#> $max_iter
#> [1] 1000
#> 
#> $accel
#> function (init, em_step, ctrl, n_x = 1, ...) 
#> {
#>     squarem_method = if (ctrl$order == 1) 
#>         3
#>     else "rre"
#>     incr_factor = n_x^(1/2)
#>     res = SQUAREM::squarem(init, em_step, control = list(K = ctrl$order, 
#>         method = squarem_method, minimize = TRUE, square = TRUE, 
#>         step.min0 = 0.5, step.max0 = 1, mstep = 4, objfn.inc = 1, 
#>         kr = 1, tol = ctrl$abstol * incr_factor, maxiter = ctrl$max_iter, 
#>         trace = FALSE, intermed = FALSE))
#>     list(ests = res$par, iters = res$fpeval, converge = res$convergence)
#> }
#> <bytecode: 0x564fd88c6a50>
#> <environment: namespace:birdie>
#> 
#> $order
#> [1] 1
#> 
#> $restart
#> [1] TRUE
#> 
#> $abstol
#> [1] 1e-06
#> 
#> $reltol
#> [1] 1e-06
#>